Split legacy tracking behaviour out
