+++ Nappou-2/src/main/java/org/atoiks/games/nappou2/pathway/TimeSlicedTrackingPathway.java
@@ -1,0 +1,77 @@
+/**
+ *  Nappou-2
+ *  Copyright (C) 2017-2019  Atoiks-Games <atoiks-games@outlook.com>
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.atoiks.games.nappou2.pathway;
+
+import org.atoiks.games.nappou2.Vector2;
+
+import org.atoiks.games.nappou2.entities.Trackable;
+
+/**
+ * Pathway that follows a trackable entity. The path it takes is linear.
+ */
+public final class TimeSlicedTrackingPathway extends TrackingPathway {
+
+    private final float scale;
+    private final float moveTime;
+    private final float delay;
+
+    private float time;
+    private boolean moving;
+
+    public TimeSlicedTrackingPathway(Trackable entity, float pathScale, float moveTime, float delay) {
+        this(Vector2.ZERO, entity, pathScale, moveTime, delay);
+    }
+
+    public TimeSlicedTrackingPathway(Vector2 pos, Trackable entity, float pathScale, float moveTime, float delay) {
+        super(entity);
+
+        this.scale = pathScale;
+        this.moveTime = moveTime;
+        this.delay = delay;
+
+        this.setPosition(pos);
+    }
+
+    public void setPosition(Vector2 pos) {
+        this.position = pos != null ? pos : Vector2.ZERO;
+        this.calculateEndpoint();
+    }
+
+    private void calculateEndpoint() {
+        // velocity = scale * (entity's position - my location) / (move time)
+        this.velocity = entity.getPosition().sub(this.position).mul(scale / moveTime);
+    }
+
+    @Override
+    public void update(final float dt) {
+        time += dt;
+        if (moving) {
+            if (time >= moveTime) {
+                moving = false;
+                time = 0;
+            } else {
+                this.position = Vector2.muladd(dt, this.velocity, this.position);
+            }
+        } else if (time >= delay) {
+            this.calculateEndpoint();
+            moving = true;
+            time = 0;
+        }
+    }
+}
