+++ src/main/java/org/atoiks/games/nappou2/pathway/TrackingPathway.java
@@ -23,64 +23,22 @@ import org.atoiks.games.nappou2.Vector2;
 import org.atoiks.games.nappou2.entities.Trackable;
 
 /**
- * Pathway that follows a trackable entity. The path it takes is linear.
+ * Subclasses of this pathway follows a trackable entity
  */
-public final class TrackingPathway implements UnboundPathway {
+public abstract class TrackingPathway implements UnboundPathway {
 
-    // Should this be a weak reference?
-    private final Trackable entity;
-    private final float scale;
-    private final float moveTime;
-    private final float delay;
+    // weak reference?
+    protected final Trackable entity;
 
-    private Vector2 velocity;
-    private Vector2 position;
+    protected Vector2 velocity = Vector2.ZERO;
+    protected Vector2 position = Vector2.ZERO;
 
-    private float time;
-    private boolean moving;
-
-    public TrackingPathway(Trackable entity, float pathScale, float moveTime, float delay) {
-        this(Vector2.ZERO, entity, pathScale, moveTime, delay);
-    }
-
-    public TrackingPathway(Vector2 pos, Trackable entity, float pathScale, float moveTime, float delay) {
+    protected TrackingPathway(Trackable entity) {
         this.entity = entity;
-        this.scale = pathScale;
-        this.moveTime = moveTime;
-        this.delay = delay;
-
-        this.setPosition(pos);
-    }
-
-    public void setPosition(Vector2 pos) {
-        this.position = pos;
-        this.calculateEndpoint();
-    }
-
-    private void calculateEndpoint() {
-        // velocity = scale * (entity's position - my location) / (move time)
-        this.velocity = entity.getPosition().sub(this.position).mul(scale / moveTime);
     }
 
     @Override
-    public Vector2 getPosition() {
+    public final Vector2 getPosition() {
         return this.position;
-    }
-
-    @Override
-    public void update(final float dt) {
-        time += dt;
-        if (moving) {
-            if (time >= moveTime) {
-                moving = false;
-                time = 0;
-            } else {
-                this.position = Vector2.muladd(dt, this.velocity, this.position);
-            }
-        } else if (time >= delay) {
-            this.calculateEndpoint();
-            moving = true;
-            time = 0;
-        }
     }
 }
